fix: support Liards dans tous les champs monétaires
- Remplace parseInt() par parseFloat() pour tous les montants monétaires
  dans useGameActions.js et toutes les vues concernées
- Ajoute step="0.1" sur tous les inputs de type="number" monétaires
  pour permettre la saisie de 0.1 (= 1 Liard), 0.5, etc.
- Les virements, prêts, salaires, prix peuvent désormais inclure des Liards

+++ src/components/views/DashboardView.js
@@ -101,7 +101,7 @@ const DashboardView = ({
   };
 
   const handleMintMoney = () => {
-    const val = parseInt(mintAmount);
+    const val = parseFloat(mintAmount);
     if (onAddTreasury && val > 0) {
       onAddTreasury(val);
       setMintAmount("");
